Chatbot/example/OpenAI-Linebot/requirements.txt
Flask==2.2.2
line-bot-sdk
openai
werkzeug==2.2.2